new component for shopping list input
